requirements.txt
python-telegram-bot==13.11
firebase-admin==5.2.0